requirements.txt
matplotlib==3.10.0
numpy==2.2.2
pandas==2.2.3
Pillow==11.1.0
plotly==5.9.0
plotly==5.24.1
seaborn==0.13.2
streamlit==1.39.0
streamlit==1.41.1
streamlit==1.37.1
